🔥 remove connextStore
