+++ modules/store/src/index.ts
@@ -1,10 +1,13 @@
-import { IAsyncStorage, IBackupService, IStoreService } from "@connext/types";
-import { Sequelize } from "sequelize/types";
+import { IAsyncStorage, IStoreService } from "@connext/types";
 
-import { ConnextStore } from "./connextStore";
-import { PisaBackupService } from "./pisaClient";
-import { StoreTypes } from "./types";
-import { WrappedAsyncStorage } from "./wrappers";
+import { storeDefaults } from "./constants";
+import { StoreService } from "./store";
+import { StoreOptions } from "./types";
+import {
+  WrappedAsyncStorage,
+  WrappedLocalStorage,
+  WrappedSequelizeStorage,
+} from "./wrappers";
 
 ////////////////////////////////////////
 // @connext/store exports
@@ -15,33 +18,59 @@ export { PisaBackupService } from "./pisaClient";
 
 export const getAsyncStore = (
   storage: IAsyncStorage,
-  backupService?: IBackupService,
+  opts: StoreOptions = {},
 ): IStoreService =>
-  new ConnextStore(
-    StoreTypes.AsyncStorage,
-    { storage: new WrappedAsyncStorage(storage) },
+  new StoreService(
+    new WrappedAsyncStorage(storage, opts.prefix),
+    opts.backupService,
+    opts.logger,
   );
 
 export const getFileStore = (
   fileDir: string,
-  backupService?: IBackupService,
+  opts: StoreOptions = {},
 ): IStoreService =>
-  new ConnextStore(StoreTypes.File, { backupService, fileDir });
+  new StoreService(
+    new WrappedSequelizeStorage(
+      opts.sequelize || `sqlite:${fileDir}/${storeDefaults.SQLITE_STORE_NAME}`,
+      opts.prefix,
+    ),
+    opts.backupService,
+    opts.logger,
+  );
 
-export const getLocalStore = (backupService?: IBackupService): IStoreService =>
-  new ConnextStore(StoreTypes.LocalStorage, { backupService });
+export const getLocalStore = (
+  opts: StoreOptions = {},
+): IStoreService =>
+  new StoreService(
+    new WrappedLocalStorage(opts.prefix),
+    opts.backupService,
+    opts.logger,
+  );
 
-export const getMemoryStore = (): IStoreService =>
-  new ConnextStore(StoreTypes.Memory);
+export const getMemoryStore = (opts: StoreOptions = {}): IStoreService =>
+  new StoreService(
+    new WrappedSequelizeStorage(
+      opts.sequelize || `sqlite:${storeDefaults.SQLITE_MEMORY_STORE_STRING}`,
+      opts.prefix,
+    ),
+    opts.backupService,
+    opts.logger,
+  );
 
 export const getPostgresStore = (
-  sequelize: Sequelize | string,
-  prefix?: string,
-  backupService?: PisaBackupService,
+  postgresUrl: string,
+  opts: StoreOptions = {},
 ): IStoreService =>
-  new ConnextStore(
-    StoreTypes.Postgres,
-    { sequelize, backupService, prefix },
+  new StoreService(
+    new WrappedSequelizeStorage(
+      opts.sequelize || postgresUrl,
+      opts.prefix,
+      opts.separator,
+      storeDefaults.DATABASE_TABLE_NAME,
+    ),
+    opts.backupService,
+    opts.logger,
   );
 
 ////////////////////////////////////////
@@ -53,7 +82,7 @@ export { storeDefaults, storeKeys, storePaths } from "./constants";
 export { PisaBackupService as PisaClientBackupAPI } from "./pisaClient";
 export { StoreTypes } from "./types";
 export {
-  KeyValueStorage,
+  StoreService,
   WrappedAsyncStorage,
   WrappedLocalStorage,
   WrappedMemoryStorage,
