+++ modules/store/src/wrappers/index.ts
@@ -1,5 +1,4 @@
 export * from "./asyncStorage";
-export * from "./keyValueStorage";
 export * from "./localStorage";
 export * from "./sequelizeStorage";
 export * from "./memoryStorage";
